refactor: 멘토링 신청 상태 관련 타입 및 상수의 경로 수정 #370
